No need to handle orgs claims specially
We handle project claims specially because some users can be in many,
many projects, and so we need to squeeze every byte that we can. But for
orgs, most users will be in just one, and those who are in more than one
are usually going to be in just two or three. So we can just do a simple
array and let the standard JSON serialization handle it.