Forbid non-project members from loading project page
Now that orgs can own projects, and anyone can see the list of projects
that an org owns, non-project members will sometimes be clicking on
project links. Changing the project table to hide links if the viewer
doesn't have permission is complicated, so for now we'll go with the
simple expedient of showing a 404 page if the user clicks on a link to a
project he's not supposed to see. We'll change this later.