Query username in org page, don't use locked or emailVerified fields
The org members table doesn't need to show the "user is locked" or
"email not verified" icons; those can be left on the admin dashboard.
But we do want the org managers to be able to see members' usernames,
since there may very well be guest users in the org with usernames but
no email addresses, and org managers should be able to help them get
Send/Receive configuration set up properly.