Org managers have full access to projects owned by their org (#919)
* Org managers have full access to projects in their org

Org managers can view, edit, and sync any projects owned by the org they
manage, even if they themselves are not explicitly listed as a member of
that project. This is similar to how site admins have full access to
anything, except restricted to projects owned by the org(s) in question.

* Await PermissionService calls that are now async

Some PermissionService calls now have to be async because they query the
database. We try to cache the results of those queries when possible to
save time, but the queries still need to be async. Async calls can't be
used in the Where clause of a LINQ expression, so CanSyncProject had to
be split into sync and async versions; the sync version doesn't query
whether a user is an org manager/admin, but the only place where it's
called is in a method that has already checked org managership earlier.

* Allow org admins to edit project data in frontend

* Fix LexAuthUserTests

* Set up to test cloning confidential projects

First step is to teach integration fixture how to create confidential
projects.

* IntegrationFixture can now create projects owned by orgs

Now we're ready to test Send/Receive of confidential projects by an org
manager who's not explicitly listed as a project member.

* Test S/R of confidential project by org manager

* Cache project ID to org IDs list for 1 hour

This will greatly reduce the number of lookups that need to be made when
an organization manager are trying to manage a project that he's not an
explicit member of.

* Also cache project confidentiality for 1 hour

---------

Co-authored-by: Kevin Hahn <kevin_hahn@sil.org>