Invalidate correct GraphQL cache
Since the org page query asks for `orgById`, the GraphQL cache is
caching the type OrgById rather than the type Organization.

While we're at it, we also invalidate the OrgById cache for a few other
mutations like adding and removing members from an org. And we stop
asking for the mutation to return the list of org members, because the
org admins don't have permission to access the .members field of orgs in
normal queries, only in the orgById query. This stops the GraphQL
permissions error that was popping up when org admins tried to do these
operations.