Extract language tags from FLEx projects and store in DB metadata (#952)
Extract language tags from FLEx projects and store them in the
FLExProjectMetadata table just like lex entry counts. Currently
we do not attempt to extract lang tags from WeSay projects.

Currently the language tags are extracted, but not analysed to try
to determine the primary language. Instead, the vernacular and analysis
tags are placed directly in the UI. The `isActive` flag means that the
tag appeared in the "Current" list, while the `isDefault` flag means
that the tag was the first one in the "Current" list. There should be
only one tag with `isDefault`.

* Add command to extract LangProject GUID as well

While we're adding hg runner commands, we can easily add one to extract
the GUID from the LangProject element, which may allow identifying
projects which started out as copies of each other even though their
Mercurial repo histories have no common commits.

* Add LangProjectId column to FlexProjectMetadata

Also include a DB migration and GQL schema update

* Make writing system ordering match what FLEx does

The order in which FLEx returns writing systems is current first, then
all non-current (if any) at the end. We want what we store in the
project metadata to match that ordering.

* Add UI for writing systems on proj page

* Add button for admins to refresh language list

* Add project controller actions to update missing langprojectid and writingsystems

---------

Co-authored-by: Kevin Hahn <kevin_hahn@sil.org>