Use GraphQL mutations for FLEx metadata updates (#973)
This changes `updateLexEntryCount` and `updateLanguageList` from being
HTTP POST queries to being GraphQL mutations, which means the GraphQL
cache is automatically updated when they run. Previously, you could
"lose" the updated entry count for a project by navigating away from it
and back to the project without a browser reload, because the result of
calling `updateLexEntryCount` was being stored on the page but not
actually sent to the GraphQL query cache, and the second time the
project loaded the GraphQL cache would serve up stale data.

Now that those two updates are using GraphQL mutations, and returning
the updated values from the mutation, the GraphQL query cache is doing
its job and caching the updated values, so navigating away from a
project and back to it works as it should.