Switch to using project page's invite modal
Instead of a new API endpoint in ProjectController, we'll just populate
the AddProjectMember modal with the user's name and ID (once it acquires
the ability to have a user ID), and then the flow will pass through the
existing GraphQL mutations. That way we won't need to do any special
work to update the GraphQL cache. It means one extra click for the admin
to approve the project join request, but it also means that admins can
choose to have the joining user join as a manager if they want, which
will come in handy sometime.