Show related projects in same org when creating project (#979)
When the user selects an org in the create-project page, any projects
with similar names in that org are shown in a list of possibly-related
projects, and the user is invited to choose one of them to join instead
of creating a new project. If he chooses to join the project, an email
is sent to that project's manager(s) to ask them to approve it. Further
down the create-project form, when a language code is entered, another
GraphQL query is made searching for any projects with a vernacular
writing system matching that language code, and those projects are added
to the related-projects list. (Ordered first, above the related-by-name
results, because they're more likely to be correct).

The related-projects list replaces the project description field and the
submit button, so the user must interact with it, either by clicking on
"Yes, join project" or by clicking on "No thanks, create new project"
before he can move on to creating the project. If he dismisses it by
clicking "No thanks", he can still change his mind and see the list of
related projects again.

When the project manager receives the email saying "(name) wants to join
the project", he can click on the button to approve it. That button will
take him to the project page, with the "Add member" modal already filled
in with the new user's name and (invisibly) GUID. He can then select a
role, either editor or manager, for the new user, and click on Add.