Allow users to accept multiple invitations
Now that we use invitation links for both projects and orgs, it will
start to become rather common for someone to receive an org invitation
and a project invitation within the same few minutes. So we need to
allow users to click on invitation links even if they already have a
Lexbox account.