Don't use GQL queries, use alternate page
Instead of GQL queries to check if the user already exists, we put the
logic into LoginController. This works: users can now receive multiple
invitation links, and only see the registration page if they have not
registered yet. If they have registered before, they're taken to an
alternate URL that logs them in immediately.

Still to do: ensure that the acceptInvitation backend doesn't throw an
error if the user is already a member of the project, but just silently
returns success if we're already in the desired state.