Better approach than changing LoginRedirect
Rather than redirect to the frontend and then query the backend to find
out if the user already exists, let's redirect to a backend endpoint,
which will then decide which frontend page to redirect to. A double
redirect may not be ideal, but it's better than special-case code in
the LoginRedirect method.