Don't throw when inviting guest users to project (#1077)
Guest users don't have email addresses, but they can be invited to
projects. So when sending the "You've been invited to project X" email,
if the user doesn't have an email address, that's not an error that
should cause an exception to be thrown. It just means that that user is
a guest user without an email, so we should just skip emailing them.