Hide members of confidential projects (#1031)
* Hide members of confidential projects

Only the following people can see the list of members if a project is
confidential: the manager(s) of the project, site admins, and the org
admin(s) of the org(s) the project belongs to, if any. Everyone else,
including people who are themselves members of the project, cannot see
its membership.

* Hide members list of confidential projects

GQL query will return a users list consisting of only one entry, the
project member currently viewing the page, if he's not allowed to see
the other project members.

* Change confidentiality default for project members

For project members, and ONLY for project members, we will default to
showing them for projects whose confidentiality has not been set. This
isn't the most secure default, but it will avoid what would likely be a
lot of user confusion.

---------

Co-authored-by: Tim Haasdyk <tim_haasdyk@sil.org>