Allow multiple project invites for one user (#1043)
* Allow users to accept multiple invitations

Now that we use invitation links for both projects and orgs, it will
start to become rather common for someone to receive an org invitation
and a project invitation within the same few minutes. So we need to
allow users to click on invitation links even if they already have a
Lexbox account.

* Forbid changing email when accepting invite

This prevents stealing someone else's invite by snooping on their email.

* Don't error if same invite accepted twice

If a user clicks the same invitation link twice, he should not get an
error message, he should just be taken to his home page. He's already a
member of the project, so trying to join the project should not error
but just succeed silently.

* Handle edge case of registering existing email

This can only happen if an admin creates the account while the user was
filling out the registration page, so this probably will never happen.