Get FLEx model version from repo and store in metadata (#1082)
Project page now shows the version number of FieldWorks that was used to
create the project, based on the model version appearing at the tip of
the hg repo. (A tooltip also shows the raw model number, e.g. 7000072,
when hovered over). This is NOT automatically updated on hg push, as we
expect this to rarely ever change.