Copy implementation from FwLiteProjectSync
This is copied nearly as-is, but with an ASP.NET Core API wrapped around
it instead of a System.CommandLine set of arguments and options. Next
step would be to move the implementing classes into a CrdtMerge.Core
library so that FwLiteProjectSync can reference them and we DRY
ourselves out again.