Instead of copying code, reference the project
This requires removing `<SelfContained>true</SelfContained>` from the
FwLiteProjectSync.csproj file, but if in the future we do decide to
deploy FwLiteProjectSync as a self-contained exe, we can just pass the
appropriate compiler flag on the command line. Meanwhile, this avoids
needing to create a CrdtSync.Core library while still sharing code.