Allow `dotnet restore LexBox.sln` to run on Linux
When running `dotnet restore` on Linux against the whole solution, it
fails with the following error message:

error NETSDK1100: To build a project targeting Windows on this operating system, set the EnableWindowsTargeting property to true.

We don't want to set EnableWindowsTargeting on Mac, so we'll use a
condition to only set it when building on Linux.