CrdtMerge API web service (#1136)
Create a new web service called CrdtMerge, which offers a single API
endpoint, `/sync?projectCode=(code)`. It will eventually take a
`projectName` parameter as well. This does a Send/Receive first to
ensure that any changes from others have been pulled into the .fwdata
file. It then synchronizes the CRDT view of the project with the .fwdata
file, and then does another Send/Receive to push the CRDT changes (now
in .fwdata) up to the project on Lexbox so that other FieldWorks users
can receive them.

This PR does not contain a Kubernetes deployment for CrdtMerge; that
will come in a follow-up PR.

---------

Co-authored-by: Kevin Hahn <kevin_hahn@sil.org>