Prep code in `src` for Svelte 5 migration (#1176)
Svelte 5 converts slots to snippets, which are passed in as props. So
if a slot had the same name as a prop, one or the other has to be
renamed to avoid a name conflict. The Svelte 5 migration tool won't
rename slots or props, so we need to do it before running the tool so
that it can migrate our code successfully.

* Prep FilterBar for Svelte 5 migration
* Prep DetailsPage for Svelte 5 migration
* Prep RadioButtonGroup for Svelte 5 migration