Have lexbox API container check fw-headless health (#1192)
Since fw-headless container isn't exposed to outside, but we might want
to know if it's healthy, we'll add code into the lexbox-api health check
to call the fw-headless health check and report lexbox-api's status as
unhealthy if fw-headless is not healthy. (Except on local dev, where
lexbox-api's status would just be Degraded so that editing fw-headless
code doesn't bring down the whole API server).