Rename ProjectSyncStatusService so it's clearer
It is now called SyncJobStatusService, to better distinguish what it
does vs. what the ProjectSyncStatus represents. The job status only
checks whether there's an active sync job running, while the project
sync status will also check how many commits are on the server waiting
to be synced. (If a sync is running right now, then the answer is 0 and
we can return without hitting the database).