Add custom component for markdown with new-tab links
That way instead of having to do a Typescript cast every time we want to
render Markdown that opens links in a new tab, we can do the cast once
and just call a different component. We've done this already for the
lexbox frontend code in src/, so we'll do the same thing for the viewer.