Add merge status API to fw-headless (#1294)
Creates a new GET /api/crdt-sync-status endpoint.

Returns 404 if the project ID does not match any project in the lexbox
database. For all other projects, this returns Syncing if a CRDT sync is
currently in progress, NeverSynced if there is no CRDT project yet, and
ReadyToSync if the CRDT project exists but is not currently syncing. If
the status is ReadyToSync, it will also return a count of commits
available in the Lexbox DB that are not yet present in the local CRDT
database. (The count is naive, just doing (lexbox commits - local
commits), so if the local DB has commits not present in Lexbox the count
could even end up negative.)