Make Node GC kick in before available RAM runs out (#1343)
Our ui container, in local dev, is currently allocated a max of 1250 MiB
of RAM. Node normally allows its heap to grow up to 2 GiB, resulting in
OOM errors in the container. Let's set Node's maximum heap to 1 GiB so
that the GC kicks in before Kubernetes kills the container.