Don't check part of speech GUIDs yet
As long as senses only have a PartOfSpeechId in them, it's hard to check
that property statelessly because we need to look up the PartOfSpeech in
order to determine whether it's predefined (and thus whether its GUID
needs to match one of the canonical GUIDs). For now, we'll skip checking
part of speech GUIDs until senses have an actual PartOfSpeech reference.