Fix complex form test - this one should fail
The Components property on a complex form can contain empty GUIDs for
the complex form entry ID, because we can infer that (it's the entry
we're looking at right now). But it cannot contain an empty GUID for the
component entry ID, because that's meaningless: which component is being
referenced is the whole point of the Components property.