Split developer docs into per-OS files (#1305)
* Split developer docs into per-OS files
* Add Linux instructions for silencing log spam

On Linux (only), Docker Desktop tends to spam the syslog with
debug-level messages about HTTP requests and responses. We should
provide instructions on how to silence that spam so Linux devs' /var
partitions don't fill up unexpectedly fast.

* Move installation summary back into main README

This removes some of the redundancy between the OS-specific files and
the main one.

We also note that the Mac instructions are likely incomplete.