Add default value for FeatureFlags column
Need to check whether specifying the default value as a C# object this
way will lead to EF Core creating the *same* list object for every row,
which could lead to bugs where adding a feature flag to one user ends up
adding it to every user. If so, we might use HasDefaultValueSql instead.