Fix feature flags in edit-user modal
Backend code was in the wrong place, so admins couldn't edit their own
user account, only other users. Also made the checkbox label clickable
as it should be, and no more need to hardcode the list of feature flags
in user.ts since we're getting it straight from the GraphQL types.