Let frontend code deal with different serialization
GraphQL serializes enums with SnakeCaseUpper naming rules, so FwLiteBeta
becomes FW_LITE_BETA. But the value stored in the backend, and thus in
the JWT, is being serialized as FwLiteBeta. To deal with this difference,
we simply do a case-insensitive comparison after removing underscores
when we're checking for the presence of a feature flag.