Add basic feature flags (#1428)
* Add feature flags to backend
* Add Svelte component for checking feature flags
* Allow admins to edit feature flags in admin dashboard

Edit-user modal has a set of checkboxes for known feature flags.

* add the feature flag to lex auth tests 
* fix feature flag serialization issues
* write a test to ensure all claims are reflected in our test user
* Get feature flags into JWT
* Let frontend code deal with different serialization

GraphQL serializes enums with SnakeCaseUpper naming rules, so FwLiteBeta
becomes FW_LITE_BETA. But the value stored in the backend, and thus in
the JWT, is being serialized as FwLiteBeta. To deal with this difference,
we simply do a case-insensitive comparison after removing underscores
when we're checking for the presence of a feature flag.

* Add DB migration for FeatureFlags column
* Fix "known good JWT" unit test

---------

Co-authored-by: Kevin Hahn <kevin_hahn@sil.org>