Improve API to query sync status (#1454)
* SendReceiveService can now count pending hg commits

This will be useful in the updated sync-status API

* Update sync-status API to include commit dates

Also includes code to check pending Mercurial commits, and report that
in the sync status as well.

* Add Lexbox API to get sync status and trigger sync
* Address review comments
* Document that -1 hg changes means "all of them"

A -1 returned in sync status means "all the hg changes on the server
will be pulled, but since there's no local clone we can't easily count
how many there will be".