Use Svelte 5 without migration (#1283)
* Update NPM packages to Svelte 5

* Stop Svelte warning about self-closing HTML tags

We'll fix the self-closing tags when we migrate to Svelte 5; in the
meantime, we don't want these warnings to flood the compilation results
and end up hiding other, more important, warnings.

* Enable preprocessing script tags in vitePreprocess

This fixes the error about TypeScript enums not being natively supported

* Avoid ".svelte-kit/tsconfig.json not found" error

By making the .svelte-kit directory (produced by svelte-kit sync) part
of the docker image in dev mode, we can avoid having the viewer build
fail the first time, which causes the container to restart unnecessarily.

* Fix up Typescript errors

* MaybePromise<T> = T | Promise<T> no longer exported from svelte-kit
* svelte/store no longer exports Invalidator type (now () => void)
* svelte-exmarkdown v4 needed for correct Svelte 5 component types
* svelte-exmarkdown renderer prop needs casting as Component<any> so
  that Typescript won't complain about incompatible prop types
* Components no longer have a .render method, instead you import the
  render function from Svelte and call it
* svelte/compiler no longer exports walk, we're supposed to use
  estree-walker instead
* estree-walker doesn't think nodes should have 'Element' type, as its
  types are designed for JS syntax trees, not Svelte component trees
* Could not get Typescript to accept our componentMap in emails.ts, as
  each component had incompatible props. Punted by declaring its type as
  Record<EmailTemplate, any>.

* Fix up Typescript errors in viewer

* Update eslint related packages and minor edits

* include pnpm-workspace as an install source for taskfile when running pnpm install

* Better type for EmailTemplate component map

* Make svelte-eslint-parser relax about self-closing tags

* Add component to encapsulate new-tab Markdown logic

This allows us to do the renderer plugin typecast just once rather than
everywhere the Markdown component is used.

* Silence lint errors that aren't really errors

* Fix all but one lint error in viewer

The remaining error is something we can't fix until we migrate our code
to runes mode, so we'll turn it into a warning in a separate commit.

* Only warn about non-reactive object props

This isn't something we can fix until we migrate our code to runes mode,
so we don't want to silence the warning but we don't want to turn it
into an error either.

* Run svelte-kit sync before pnpm run -r build

Some parts of the build were erroring out due to the .svelte-kit
directory not being present. This should create that directory before
the parts of the build that need it.

* Simplify handling of EmailTemplateProps

* Add custom component for markdown with new-tab links

That way instead of having to do a Typescript cast every time we want to
render Markdown that opens links in a new tab, we can do the cast once
and just call a different component. We've done this already for the
lexbox frontend code in src/, so we'll do the same thing for the viewer.

* Probably revert this and just replace happy-dom

* Fix tests, svelte-check and lint

* tidy up and maybe fix build

* Ensure verbatimModuleSyntax is enabled

* Fix docker build

* Typeahead allows selecting with keyboard (#1462)

* Typeahead allows selecting with keyboard
* Make typeahead results close on enter and fixup event handler
* Fix typeahead selected-user persisted after more typing
* Prevent PlainInput from emitting input events without a real input event being triggered.
* Prevent old typeahead results and queries when selecting a result

---------

Co-authored-by: Tim Haasdyk <tim_haasdyk@sil.org>

* Bump PNPM to 9.15.5 (#1440)

* Bump PNPM to 9.15.5

PNPM v9.15.5 brings in several bugfixes, at least one of which (running
preprepare scripts from packages on pnpm install, and "preprepare" is
not a typo) we'll probably want to have when we upgrade our SvelteKit
dependency.

* Prep for PNPM 10: add "prepare": "svelte-kit sync"
* Work around corepack bug preventing PNPM install

* Improve API to query sync status (#1454)

* SendReceiveService can now count pending hg commits

This will be useful in the updated sync-status API

* Update sync-status API to include commit dates

Also includes code to check pending Mercurial commits, and report that
in the sync status as well.

* Add Lexbox API to get sync status and trigger sync
* Address review comments
* Document that -1 hg changes means "all of them"

A -1 returned in sync status means "all the hg changes on the server
will be pulled, but since there's no local clone we can't easily count
how many there will be".

* Disable hg bundle2 compression (#1465)

* Allow all host names for vite dev server

* Fix FW Lite for Svelte 5

* Remove forgotten generated enum value

* Stop using the "FLEx" abbreviation in the UI (#1473)

* Fix svelte-check and eslint

---------

Co-authored-by: Tim Haasdyk <tim_haasdyk@sil.org>
Co-authored-by: Kevin Hahn <kevin_hahn@sil.org>