Skip Send/Receive steps if no pending changes
Incoming changes are detected by using PendingCommitCount; if it's 0
then there are no Mercurial changes sitting in Lexbox, so a Send/Receive
would do nothing and we can save some time by skipping it.

Outgoing changes are detected by looking at the FwdataChanges result
from the CRDT sync code. If that is 0 then the Send/Receive would not
push any new commits so we can safely skip it and save more time.