Result of svelte-5 migration script
This is what `pnpx sv migrate svelte-5` produced, with no changes. There
is one file, `src/routes/(unauthenticated)/login/+page.svelte`, which
might need further attention: a `@migration task` commit was created
saying "review uses of `navigating`" (which used to be imported from
$app/stores and now comes from $app/state instead).

There might not be any changes needed, but if there are, that work will
go in a separate commit.