Just use `$bindable()`, not `$bindable(undefined)`
In Svelte 5, if you have set a fallback value for a bindable prop then
you're not allowed to have the caller pass `undefined` to it. But if the
"fallback value" was just `undefined` in the first place, then that's
the default behavior of just plain `$bindable()` with no fallback value.