Fix PasswordStrengthMeter without infinite loop
Instead of setting a bindable prop based on a derived from the props
list, which was triggering an infinite loop, we will fall back to
function props and pass an onScoreUpdated function into the password
strength meter component. It's slightly more verbose, but it avoids the
infinite loop of effects triggering updates to themselves that our
previous approach was causing.