Migrate Lexbox UI to Svelte 5 (#1634)
* Rename "derived" uses that mess up migration script

The `pnpx sv migrate svelte-5` script has trouble with changing `$:`
statements to `derived` when there are existing uses of `derived` in the
code, so we'll rename the three cases that have that problem before
running the migration script.

* Result of svelte-5 migration script

This is what `pnpx sv migrate svelte-5` produced, with no changes. There
is one file, `src/routes/(unauthenticated)/login/+page.svelte`, which
might need further attention: a `@migration task` commit was created
saying "review uses of `navigating`" (which used to be imported from
$app/stores and now comes from $app/state instead).

There might not be any changes needed, but if there are, that work will
go in a separate commit.

* Navigating state is guaranteed non-null

In SvelteKit version 2.12 or later, the `navigating` object is
guaranteed never to be null; its properties (from, to, etc.) will be
null if no navigation is going on, but you can safely check
`navigating.to` rather than `navigating?.to`.

* Fix now-incorrect ordering of variables

IconButton was using the `size` prop before it was declared. In Svelte 4
the compiler was reordering this, in Svelte 5 this is no longer allowed;
the variable that uses `size` needs to be moved after the $props call.

* Just use `$bindable()`, not `$bindable(undefined)`

In Svelte 5, if you have set a fallback value for a bindable prop then
you're not allowed to have the caller pass `undefined` to it. But if the
"fallback value" was just `undefined` in the first place, then that's
the default behavior of just plain `$bindable()` with no fallback value.

* Use `page` from `$app/state`, not `$app/stores`

Running the `pnpx sv migrate svelte-5` a second time allowed it to
migrate several uses of the `$page` store into the new `page` state.

This commit is the unchanged result of the migration tool.

* Allow eslint to fix auto-fixable errors

These are errors that used to be suppressed but are no longer errors,
and eslint automatically removed the error-suppressing comments.

* Fix consistent-type-imports errors

Types like `import('svelte').Snippet` are now forbidden, so we replace
them with explicit type imports at the top of the file.

Prettier made other code changes when I saved these files, which I've
placed into a separate commit to make refactoring slightly easier should
that prove necessary.

* Changes made by Prettier when fixing lint errors

Separated into separate commit so these changes will be slightly easier
to revert if they are wrong.

* Fix remaining lint errors
* Fix type errors
* Remove now-unused createBubbler() imports
* change custom overlay to use passed callbacks rather than custom events
* Pass rest props through intermediate Select components
* Fix errors on project page
* Address review comments
* Get rid of infinite loop on register page
* Fix PasswordStrengthMeter without infinite loop

Instead of setting a bindable prop based on a derived from the props
list, which was triggering an infinite loop, we will fall back to
function props and pass an onScoreUpdated function into the password
strength meter component. It's slightly more verbose, but it avoids the
infinite loop of effects triggering updates to themselves that our
previous approach was causing.

* Avoid infinite recursion during SSR

This works just fine in client-side rendering, but causes an infinite
recursive call to headerContent() when rendered server-side. Putting a
`$derived` in the middle breaks the loop and makes the page render
correctly in both SSR and CSR.

---------

Co-authored-by: Kevin Hahn <kevin_hahn@sil.org>
Co-authored-by: Tim Haasdyk <tim_haasdyk@sil.org>