Remove createEventDispatcher from OrgTabs
The version using `dispatch()` checked a return value to see if the
event was canceled using event.preventDefault(), and only changed tabs
if the event was not canceled. Prop functions don't use preventDefault
since they don't use DOM events, so I changed it so that this method can
return a boolean. The `dispatch` logic was "false means it was canceled"
but I'm using true to mean canceled so that if you don't care about
canceling, you can simply return nothing and a falsy value will allow
tab navigation to proceed. This inverts the meaning of the boolean, but
since nobody was using it it doesn't really matter.