Fix infinite effect loop in EditUserAccount
Accessing the `$form` in onScoreUpdated was causing Svelte to track the
entire form as a dependency of the PasswordStrengthMeter component,
causing an infinite effect loop as soon as we updated `$form.score`.
Svelte is smart enough not to track dependencies for `$form.foo = bar`
but needed to be told not to track the `if ($form)` part.