Followup work on Svelte 5 migration - code cleanup and so on (#1652)
* Get rid of recursive updates in run() calls

TrainTracks and HgLogView components have slightly tricky logic that the
migration tool didn't quite figure out correctly. Pulling the run() call
apart into multiple `$derived` sections is much cleaner.

* Change `$props()` to const if not bindable

* Convert on:change in FilterBar to prop function

Also got to simplify a convoluted mess around ViewMode in the default
authenticated home page. Now that `$derived` values can be overridden,
it turned into a very simple `$derived(cookieVal ?? defaultVal)` line,
while still allowing the user to set the value.

* Remove most uses of `run()` from svelte/legacy

A couple uses of `run()` need something non-trivial done to replace
them. In one case we need to write better logic in lexSuperForm: the
project code should probably be split into two form entries, one
calculated code and one override, and the checkbox when checked smoothly
replaces the calculated code field with the override field.

In the other case where I kept a `run()` call, we will be able to
replace it easily once async deriveds become available in Svelte 5.
(They're available now in experimental mode, but it's been less than a
week so I want to wait until they settle a bit more).

* Remove more `run()` usages

* Remove `run()` use in project creation page

* Remove a preventDefault use from svelte/legacy

* Remove redundant undefineds in props

Optional props don't need `| undefined` in their type definition, as
it's implied by the `?`. And the assignment via `$props` likewise
doesn't need to set a default `undefined` value to optional props.

* Remove createEventDispatcher from components

* Get rid of `run()` in FilterBar and UserTypeahead

Also replaced PlainInput with a DebouncedInput component, so that we can
switch PlainInput back to being what it sounds like: a plain input field
that doesn't try to do anything clever.

* Remove `run()` usage from email tester page

Note that this would be better as an async derived once it's available
(see discussion #15845 in the Svelte repo for details). Until then,
this works well enough.

* Make PlainInput plain again

* UserTypeahead now uses a PlainInput

* Fix infinite effect loop in EditUserAccount

Accessing the `$form` in onScoreUpdated was causing Svelte to track the
entire form as a dependency of the PasswordStrengthMeter component,
causing an infinite effect loop as soon as we updated `$form.score`.
Svelte is smart enough not to track dependencies for `$form.foo = bar`
but needed to be told not to track the `if ($form)` part.

* Fix customCode reactivity

* Make code-exists error initially undefined

* Remove debouncing prop and refactor DebouncedInput to accept values pushes

* Fix code_exists error shown delayed

---------

Co-authored-by: Tim Haasdyk <tim_haasdyk@sil.org>