Fix "file being used by another process" errors
When you POST a file and then PUT a replacement soon afterwards, as in
our integration tests, it can result in the error "Access denied because
file is being used by another process". This is a Windows-ism that is
not necessary in a Linux filesystem, but the File.Open method by default
uses exclusive locking because of how Windows filesystems work. So we
have to explicitly request no locking when we write the file.