Fix loading FileMetadata from form
Turns out that when you specify [FromForm] on a complex object, ASP.NET
understands that to mean "the entire form is this object" and loads the
object's properties by name from the form fields. That explains why we
were getting FileMetadata objects with only the filename filled in. Now
we allow that (so that client code can create forms with author, license,
etc. as separate fields) but also parse JSON from a "metadata" field. If
both are used, they are merged.