Fix inverted newFile logic from refactor
The refactor returns a newFile argument, but didn't change the existing
logic that was looking for replacingExistingFile (i.e. the opposite of
newFile). So the HTTP result was Created when it should have been OK and
vice-versa. Also cleaned up a couple Console.WriteLine leftovers, and
added the projectId form entry in the tests since it's now required for
all operations, even if fileId is specified.