File upload and download API (#1744)
All API endpoints are in /api/media, as follows:

* POST /api/media/?projectId=(guid)

Uploads a file, returning 201 Created and a Location header where that
file can be found. Project ID is required since files are stored in the
project folder (Mercurial repo) of FieldWorks projects. The body of the
HTTP 201 Created response will also contain the file ID (a GUID) which
could be used in subsequent GET or PUT requests.

Optionally, a `fileId` form value can be included along with the upload
request, in which case this will be treated just like a PUT request to
replace a previously-uploaded file.

Any desired metadata (author, license, etc.) can be included as form
values alongside the uploaded file.

* PUT /api/media/(guid)

Replaces an existing file; the GUID in the URL is the one previously
returned from a POST request (or the URL can be extracted from the
Location header). A `projectId` form value (containing a GUID) is
required, and an error will be returned if the projectId does not match
the one the file is currently in (i.e. you cannot move files to other
projects via this API).

Similar to POST, metadata can be included as form values.

* GET /api/media/(guid)

Downloads a file previously uploaded. Supports Range header so that
interrupted downloads can be resumed. Also supports If-None-Match header
using the SHA256 checksum of the file, to allow skipping a download if
the client already has an uncorrupted copy of the file.

* DELETE /api/media/(guid)

Deletes the file.

* GET /api/metadata/(guid)

Returns metadata about the file: filename, size, SHA256 checksum, and
any other metadata uploaded via PUT or POST earlier.

---------

Co-authored-by: Kevin Hahn <kevin_hahn@sil.org>