Roll over log files based on file size (#1781)
File size and max log count can be set in configuration, default to max
of 2 files of 50 megabytes each so that if config not set we don't end
up using up too much space for logs.

Also set up sharing log files so that if the logs have rolled over at
least once, it will grab the latest *and* previous log files to share,
so that we don't end up in a situation where the log rollover just
happened and we shared a log file with only 5 lines in it.