Backend now knows if user can download any project
Site admins and org admins can download arbitrary projects without
needing to be a member (projects owned by their org, or any project at
all in the case of site admins), so let's return a boolean flag for the
frontend to enable UI for that. The UI will still need to handle "Hey,
you're not actually authorized to download that project" because it will
be used by org admins, not only by site admins who have all access.